decimales

+++ data/egresos/egresos.js
@@ -363,7 +363,7 @@ function comprobar2() {
                                         }
                                         if ($("#cantidad_unidad").val() != "")
                                         {
-                                            cantidad_unidad = parseInt($("#cantidad_unidad").val()) * parseInt($("#cantidad").val());
+                                            cantidad_unidad = parseFloat($("#cantidad_unidad").val()) * parseFloat($("#cantidad").val());
                                             unidad_medida = $("#unidad_medida")[0].selectedOptions[0].text;
                                             unidad_medida = unidad_medida.split("--");
                                             unidad_medida = unidad_medida[0];
@@ -426,7 +426,7 @@ function comprobar2() {
                                                 }
                                                 if ($("#cantidad_unidad").val() != "")
                                                 {
-                                                    cantidad_unidad = parseInt($("#cantidad_unidad").val()) * parseInt($("#cantidad").val());
+                                                    cantidad_unidad = parseFloat($("#cantidad_unidad").val()) * parseFloat($("#cantidad").val());
                                                     unidad_medida = $("#unidad_medida")[0].selectedOptions[0].text;
                                                     unidad_medida = unidad_medida.split("--");
                                                     unidad_medida = unidad_medida[0];
@@ -475,7 +475,7 @@ function comprobar2() {
                                             }
                                             if ($("#cantidad_unidad").val() != "")
                                             {
-                                                cantidad_unidad = parseInt($("#cantidad_unidad").val()) * parseInt($("#cantidad").val());
+                                                cantidad_unidad = parseFloat($("#cantidad_unidad").val()) * parseFloat($("#cantidad").val());
                                                 unidad_medida = $("#unidad_medida")[0].selectedOptions[0].text;
                                                 unidad_medida = unidad_medida.split("--");
                                                 unidad_medida = unidad_medida[0];
